--- src/sys/p2/time/utils.rs
@@ -1,5 +1,0 @@
-use std::io;
-
-pub(crate) fn timeout_err(msg: &'static str) -> io::Error {
-    io::Error::new(io::ErrorKind::TimedOut, msg)
-}
